using array length types

+++ week-2/ex-15.c
@@ -2,10 +2,10 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-const int ARRAY_SIZE = 20;
+const size_t ARRAY_SIZE = 20;
 const int MIN_NUMBER = 100;
 const int MAX_NUMBER = 120;
-const int AMOUNT_POSSIBLE_NUMBERS = MAX_NUMBER - MIN_NUMBER + 1;
+const size_t AMOUNT_POSSIBLE_NUMBERS = MAX_NUMBER - MIN_NUMBER + 1;
 
 /**
  * M = size of random num array
@@ -13,9 +13,9 @@ const int AMOUNT_POSSIBLE_NUMBERS = MAX_NUMBER - MIN_NUMBER + 1;
  * O(M+N) runtime implementation
  * O(M+N) Memory
  */
-void fill_and_check_occurrence(int * array, int * occurrence_map, int size, int min, int max) {
+void fill_and_check_occurrence(int * array, int * occurrence_map, size_t size, int min, int max) {
     srand(time(NULL));
-    int i; int ran_num;
+    size_t i; int ran_num;
     for (i = 0; i < size; i++) {
         ran_num = min + (rand() % (max + 1 - min));
         array[i] = ran_num;
@@ -30,7 +30,7 @@ int main() {
     fill_and_check_occurrence(random_numbers, occurrence_map, ARRAY_SIZE, MIN_NUMBER, MAX_NUMBER);
 
     printf("All random numbers: ");
-    int i;
+    size_t i;
     for (i = 0; i < ARRAY_SIZE; i++) {
         printf("%d ", random_numbers[i]);
     }
@@ -38,7 +38,7 @@ int main() {
     printf("\nAll missing numbers: ");
     for (i = 0; i < AMOUNT_POSSIBLE_NUMBERS; i++) {
         if (occurrence_map[i] == 0) {
-            printf("%d ", i + MIN_NUMBER);
+            printf("%d ", (int) i + MIN_NUMBER);
         }
     }
 }
